rofl_app/requirements.txt
setuptools>=68.0.0
web3==6.14.0
httpx>=0.24.0
eth-account==0.10.0
eth-abi==5.0.1
eth-utils==3.0.0
typing-extensions>=4.5.0
python-dotenv==1.0.1
requests==2.31.0
pyyaml>=6.0.1
asyncio==3.4.3
aiohttp==3.9.1
jsonschema==4.21.1
loguru==0.7.2
cryptography==42.0.2
